fix(auth): fixed bugs and refactored the old codes, removed userdetails and userdetailsservice(unnecessary)
