+++ workzone/src/main/java/com/project/workzone/repository/RoleRepository.java
@@ -4,5 +4,5 @@ import com.project.workzone.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<Role, Long> {
-    public Role findRoleById(Long id);
+    Role findRoleById(Long id);
 }
