+++ workzone/src/main/java/com/project/workzone/service/AuthService.java
@@ -1,0 +1,10 @@
+package com.project.workzone.service;
+
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.http.ResponseEntity;
+
+public interface AuthService {
+    ResponseEntity<?> loginUser(HttpServletResponse response);
+    ResponseEntity<?> updateAccessToken();
+    ResponseEntity<?> deleteTokens(HttpServletResponse response);
+}
